Arreglo de errores post-test

+++ microservice.viajes/src/main/java/com/microservice/viajes/dto/ViajeConID_DTO.java
@@ -23,7 +23,7 @@ public class ViajeConID_DTO {
 
     private int usuario_id;
 
-    private int monopatin_id;
+    private Long monopatin_id;
 
     private int tiempo;
